build: rename describe.py to descrive_apis.py
